+++ src/gamspy/__init__.py
@@ -20,6 +20,7 @@ from gamspy._algebra import (
 from gamspy._algebra.expression import Expression
 from gamspy._backend.engine import EngineClient
 from gamspy._backend.neos import NeosClient
+from gamspy._config import get_option, set_options
 from gamspy._container import Container
 from gamspy._model import Model, ModelStatus, Problem, Sense, SolveStatus
 from gamspy._options import ModelInstanceOptions, Options
@@ -73,4 +74,6 @@ __all__ = [
     "utils",
     "SpecialValues",
     "__version__",
+    "get_option",
+    "set_options",
 ]