+++ tests/test_gamspy.py
@@ -7,14 +7,31 @@ import os
 import shutil
 from os.path import join
 
+import gamspy_base
 import pytest
 
-import gamspy
+import gamspy as gp
+from gamspy.exceptions import GamspyException, ValidationError
 
 
 @pytest.mark.unit
 def test_version():
-    assert gamspy.__version__ == "1.5.0"
+    assert gp.__version__ == "1.5.0"
+
+
+@pytest.mark.unit
+def test_config():
+    m = gp.Container()
+    assert m.system_directory == gamspy_base.directory
+
+    i = gp.Set(m, records=["i1", "i2"])
+    a = gp.Parameter(m, domain=i, records=[("i1", 1), ("i2", 2)])
+    with pytest.raises(ValidationError):  # GAMSPy catches the domain violation
+        a["i3"] = 5
+
+    gp.set_options({"DOMAIN_VALIDATION": 0})
+    with pytest.raises(GamspyException):  # GAMS returns domain violation error
+        a["i3"] = 5
 
 
 @pytest.mark.doc
@@ -24,7 +41,7 @@ def test_switcher():
     with open(join(root, "docs", "_static", "switcher.json")) as file:
         switcher = json.loads(file.read())
         versions = [elem["version"] for elem in switcher]
-        assert f"v{gamspy.__version__}" in versions
+        assert f"v{gp.__version__}" in versions
 
 
 @pytest.fixture